# SWAGGER-NVSL-APIS.yaml
openapi: 3.0.3
info:
  title: NVSL API
  version: 1.0.0
  description: |
    Documentação consolidada das APIs do NVSL.
    Base URL padrão: `http://localhost:8081/api`.

    Rotas **web** na raiz da aplicação (`http://localhost:8081`, cadastro MGI):
    - `GET /` — status da API
    - `GET /redirect-gov` — callback OAuth2/OIDC (redirect_uri)
    - `GET /logout` — redireciona ao login do frontend (`?from=logout`); revogação do token: `POST /api/auth/logout`
servers:
  - url: http://localhost:8081/api
    description: Ambiente local (Docker)
...
paths:
...
  /user/perfis-ativos:
    get:
      tags: [Contexto]
      summary: Lista perfis ativos e vigentes do usuário
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de perfis ativos no contexto do usuário
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerfisAtivosResponse"
        "401":
          description: Não autenticado
  /user/trocar-contexto:
    post:
      tags: [Contexto]
      summary: Troca o perfil de contexto ativo do usuário
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrocarContextoRequest"
      responses:
        "200":
          description: Contexto alterado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrocarContextoResponse"
        "401":
          description: Não autenticado
        "403":
          description: Perfil não permitido para o usuário
        "422":
          description: Erro de validação
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
...
    UserSafe:
      type: object
      properties:
        id: { type: integer }
        govbr_sub: { type: string }
        name: { type: string }
        email: { type: string, nullable: true }
        role: { type: string }
        esfera_atuacao: { type: string }
        uf_lotacao: { type: string, nullable: true }
        municipio_lotacao: { type: string, nullable: true }
...
    TrocarContextoRequest:
      type: object
      required: [perfil_usuario_id]
      properties:
        perfil_usuario_id:
          type: integer
          example: 12
    TrocarContextoResponse:
      type: object
      properties:
        message:
          type: string
          example: Contexto alterado com sucesso.
        user:
          $ref: "#/components/schemas/UserSafe"
    PerfilAtivoItem:
      type: object
      properties:
        perfil_usuario_id: { type: integer }
        perfil_id: { type: integer }
        nome: { type: string }
        esfera:
          type: string
          enum: [federal, estadual, municipal]
        uf: { type: string, nullable: true }
        municipio: { type: string, nullable: true }
        orgao: { type: string, nullable: true }
        data_inicio_vigencia: { type: string, format: date, nullable: true }
        data_fim_vigencia: { type: string, format: date, nullable: true }
        ativo: { type: boolean }
    PerfisAtivosResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PerfilAtivoItem"
        perfil_usuario_ativo_id:
          type: integer
          nullable: true
...